requirements.txt
asgiref==3.4.1
attrs==21.2.0
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.1.9
django-cors-headers==3.7.0
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-filter==2.4.0
django-rest-resetpassword==0.1.2
django-simple-history==3.0.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
drf-spectacular==0.17.2
idna==2.10
inflection==0.5.1
ipinfo==3.0.0
ipinfo-django==1.1.0
jsonschema==3.2.0
nepali-datetime==1.0.7
Pillow==8.3.0
psycopg2==2.9.1
PyJWT==2.1.0
pyrsistent==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
